Split off Player class to separate file.
